Zadanie 7.1

+++ kodilla-stream/src/test/java/com/kodilla/stream/portfolio/BoardTestSuite.java
@@ -152,11 +152,9 @@ public class BoardTestSuite {
                 .filter(inProgressTasks::contains)
                 .flatMap(a -> a.getTasks().stream())
                 .map(b -> ChronoUnit.DAYS.between(b.getCreated(), LocalDate.now()))
-                .collect(Collectors.averagingLong(c -> c));
+                .collect(Collectors.averagingDouble(c -> c));
 
         //Then
         Assert.assertEquals(10, average, 0);
-
-
     }
 }
